# feat(mpc): mpc reformulation for faster computation