--- a/control/mpc_lateral_controller/include/mpc_lateral_controller/mpc_lateral_controller.hpp
+++ b/control/mpc_lateral_controller/include/mpc_lateral_controller/mpc_lateral_controller.hpp
@@ -52,7 +52,7 @@ using tier4_debug_msgs::msg::Float32Stamped;
 class MpcLateralController : public trajectory_follower::LateralControllerBase
 {
 public:
-  explicit MpcLateralController(rclcpp::Node & node);
+  explicit MpcLateralController(rclcpp::Node * node);
   virtual ~MpcLateralController();
 
 private: