… control/mpc_lateral_controller/CMakeLists.txt
… control/trajectory_follower_node/src/controller_node.cpp