## Description

Revise the mpc formulation for faster computation based on the proposal: https://github.com/orgs/autowarefoundation/discussions/3681

But, it makes it slower. (The original implementation is fast enough.)

The time required in the QP optimization process is below:

**Before (0.5ms ~ 1.0ms)**
![image_before](https://github.com/autowarefoundation/autoware.universe/assets/21360593/ab63f934-8d84-4e82-98ea-ca5e8b911f9d)


**After (1.0ms ~ 2.0ms)**
![image_after](https://github.com/autowarefoundation/autoware.universe/assets/21360593/1df3dcad-b653-4f0a-b844-35932ea67051)


## Related links

<!-- Write the links related to this PR. Private links should be clearly marked as private, for example, '[FOO COMPANY INTERNAL LINK](https://example.com)'. -->

## Tests performed

<!-- Describe how you have tested this PR. -->

## Notes for reviewers

<!-- Write additional information if necessary. It should be written if there are related PRs that should be merged at the same time. -->

## Interface changes

<!-- Describe any changed interfaces, such as topics, services, or parameters. -->

## Effects on system behavior

<!-- Describe how this PR affects the system behavior. -->

## Pre-review checklist for the PR author

The PR author **must** check the checkboxes below when creating the PR.

- [ ] I've confirmed the [contribution guidelines].
- [ ] The PR follows the [pull request guidelines].

## In-review checklist for the PR reviewers

The PR reviewers **must** check the checkboxes below before approval.

- [ ] The PR follows the [pull request guidelines].
- [ ] The PR has been properly tested.
- [ ] The PR has been reviewed by the code owners.

## Post-review checklist for the PR author

The PR author **must** check the checkboxes below before merging.

- [ ] There are no open discussions or they are tracked via tickets.
- [ ] The PR is ready for merge.

After all checkboxes are checked, anyone who has write access can merge the PR.

[contribution guidelines]: https://autowarefoundation.github.io/autoware-documentation/main/contributing/
[pull request guidelines]: https://autowarefoundation.github.io/autoware-documentation/main/contributing/pull-request-guidelines/