@TakaHoribe 2023-09-12
This change could improve the computational stability, but it slightly increases the computational time. I close this PR for now.